% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaged_quantile.R
\name{averaged_median}
\alias{averaged_median}
\title{Calculate the averaged (smoothed) median}
\usage{
averaged_median(
  x,
  w = NULL,
  na.rm = TRUE,
  quantiles_n = 9L,
  quantiles_w = c(1:4, 5, 4:1)
)
}
\arguments{
\item{x}{numeric vector or an R object}

\item{w}{numeric vector of sample weights the same length as x giving the weights to use for elements of x}

\item{na.rm}{logical; if true, any NA or NaN's are removed from x before computation}

\item{quantiles_n}{integer number of quantiles used for averaging; must be odd}

\item{quantiles_w}{weights used for average quantiles; length must equal quantiles_n}
}
\value{
a scalar
}
\description{
Calculate the averaged (smoothed) median
}
\examples{
averaged_median(x = mtcars$mpg)
}
