% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaged_quantile.R
\name{averaged_quantile}
\alias{averaged_quantile}
\title{Calculate the averaged (smoothed) quantile}
\usage{
averaged_quantile(
  x,
  w = NULL,
  probs = 0.5,
  na.rm = TRUE,
  quantiles_n = 9L,
  quantiles_w = c(1:4, 5, 4:1)
)
}
\arguments{
\item{x}{numeric vector or an R object}

\item{w}{numeric vector of sample weights the same length as x giving the weights to use for elements of x}

\item{probs}{numeric; percentile with value \verb{[0,1]}}

\item{na.rm}{logical; if true, any NA or NaN's are removed from x before computation}

\item{quantiles_n}{integer number of quantiles used for averaging; must be odd}

\item{quantiles_w}{weights used for average quantiles; length must equal quantiles_n}
}
\value{
a numeric vector with length probs
}
\description{
Calculate the averaged (smoothed) quantile
}
\examples{
averaged_quantile(x = mtcars$mpg, probs = c(0.25, 0.5, 0.75))
}
