% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{binipolate}
\alias{binipolate}
\title{Binipolate
Calculates binned interpolated percentiles}
\usage{
binipolate(data, x, probs = 0.5, bin_size, .by = NULL, w = NULL)
}
\arguments{
\item{data}{data frame, data frame extension (e.g. a tibble), or a lazy data frame (dbplyr, dtplyr) inherited from tidyverse}

\item{x}{column to compute}

\item{probs}{numeric vector of percentiles with values \verb{[0,1]}}

\item{bin_size}{size of binning}

\item{.by}{optional, a tidy-selection of columns for single-operation grouping}

\item{w}{numeric vector of weights the same length as x giving the weights to use for elements of x}
}
\value{
a tibble or data frame
}
\description{
Binipolate
Calculates binned interpolated percentiles
}
\examples{
binipolate(mtcars, x = cyl, bin_size = 0.25)
mtcars |> binipolate(disp, probs = c(0.25, 0.5, 0.75), bin_size = 0.25)
mtcars |> binipolate(disp, probs = c(0.25, 0.5, 0.75), bin_size = 0.25, .by = cyl, w = mpg)
}
