% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Cross-tabulate one or two variables}
\usage{
crosstab(data, ..., w = NULL, percent = NULL)
}
\arguments{
\item{data}{a data frame}

\item{...}{one or two variables, for a one- or two-way cross-tabulation}

\item{w}{weight}

\item{percent}{for a two-way cross-tabulation, replace counts with row or column percentages
\itemize{
\item NULL, the default, shows counts rather than percentages
\item "row" will replace counts with row percentages, summing to 100\% across columns within each row
\item "column" will replace counts with column percentages, adding to 100\% across rows within each column
}}
}
\value{
a tibble
}
\description{
Cross-tabulate one or two variables
}
\examples{
crosstab(mtcars, cyl)
crosstab(mtcars, cyl, gear)
crosstab(mtcars, cyl, gear, w = mpg, percent = "column")
}
