% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidatatools-package.R
\docType{package}
\name{epidatatools-package}
\alias{epidatatools}
\alias{epidatatools-package}
\title{epidatatools: Data tools used at the Economic Policy Institute}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools used by the Economic Policy Institute.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://economic.github.io/epidatatools/}
  \item \url{https://github.com/Economic/epidatatools}
  \item Report bugs at \url{https://github.com/Economic/epidatatools/issues}
}

}
\author{
\strong{Maintainer}: Ben Zipperer \email{benzipperer@gmail.com}

Other contributors:
\itemize{
  \item Jori Kandra \email{jkandra@epi.org} [contributor]
  \item Economic Policy Institute [copyright holder, funder]
}

}
\keyword{internal}
