% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls.R
\name{get_bls}
\alias{get_bls}
\title{Retrieve data from the BLS API}
\usage{
get_bls(series, start, end, metadata = F)
}
\arguments{
\item{series}{BLS series code}

\item{start}{Start year (numeric)}

\item{end}{End year (numeric)}

\item{metadata}{Flag for additional metadata}
}
\value{
a tibble
}
\description{
This function is simply a wrapper around \href{https://github.com/groditi/blsR}{\code{blsR}} and requires you to have an \href{https://www.bls.gov/developers/}{BLS API key} saved in the \code{BLS_API_KEY} environment variable.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_bls("LNU02300060", start = 2020, end = 2024)

bls_series_ids = c(
  "LNU02300060",
  emp_fb_2534 = "LNU02073399",
  epop_asianmen_2554 = "LNU02332330Q",
  injuries_annual = "ISU00000000031004",
  cpi_semi = "CUUS0000SA0"
)

get_bls(bls_series_ids, start = 2025, end = 2025)

# Tibble with list columns including additional metadata
complete_results = get_bls(bls_series_ids, start = 2025, end = 2025, metadata = T)
complete_results

complete_results |>
  tidyr::unnest(metadata)
\dontshow{\}) # examplesIf}
}
