% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{interpolated_median}
\alias{interpolated_median}
\title{interpolated_median: binned interpolated median}
\usage{
interpolated_median(x, bin_size, w = NULL, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
interpolated_median: binned interpolated median
}
