% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{interpolated_quantile_p}
\alias{interpolated_quantile_p}
\title{interpolated_quantile helper}
\usage{
interpolated_quantile_p(data, p)
}
\arguments{
\item{p}{}
}
\description{
interpolated_quantile helper
}
