% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_micro.R
\name{dl_ipums_micro}
\alias{dl_ipums_micro}
\alias{dl_ipums_acs1}
\alias{dl_ipums_asec}
\alias{dl_ipums_cps}
\title{Download a selection of IPUMS microdata extracts}
\usage{
dl_ipums_micro(extract)

dl_ipums_acs1(years = NULL, variables, description = NULL, ...)

dl_ipums_asec(years = NULL, variables, description = NULL, ...)

dl_ipums_cps(months = NULL, variables, description = NULL, ...)
}
\arguments{
\item{extract}{an IPUMS microdata extract as defined by \code{\link[ipumsr:define_extract_micro]{ipumsr::define_extract_micro()}}}

\item{years}{a vector of years}

\item{variables}{a vector of variable names, or a list of detailed variable specifications as created by \code{\link[ipumsr:var_spec]{ipumsr::var_spec()}}}

\item{description}{description for the extract}

\item{...}{arguments passed to \code{\link[ipumsr:define_extract_micro]{ipumsr::define_extract_micro()}} other than collection, description, samples, variables}

\item{months}{a vector of dates representing months of CPS samples.}
}
\value{
a tibble of microdata from the IPUMS API
}
\description{
Convenience functions for downloading samples and variables from \href{https://www.ipums.org/}{IPUMS} microdata using their \href{https://developer.ipums.org/}{API} and package \href{https://tech.popdata.org/ipumsr/articles/ipums-api-micro.html}{\code{ipumsr}}.
}
\details{
These functions are simply wrappers around \href{https://tech.popdata.org/ipumsr/}{\code{ipumsr}} and require you to have an \href{https://tech.popdata.org/ipumsr/articles/ipums-api.html#set-key}{IPUMS API key} saved in the \code{IPUMS_API_KEY} environment variable.
}
\section{Functions}{
\itemize{
\item \code{dl_ipums_micro()}: base function group

\item \code{dl_ipums_acs1()}: Download IPUMS ACS 1-year files

\item \code{dl_ipums_asec()}: Download IPUMS CPS ASEC

\item \code{dl_ipums_cps()}: Download IPUMS Monthly CPS

}}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example ASEC download
dl_ipums_asec(2021:2023, c("YEAR", "OFFPOV", "ASECWT"))

# example monthly CPS download
begin_month = lubridate::ym("2022 September")
end_month = lubridate::ym("2024 August")
cps_months = seq(begin_month, end_month, by = "month")
dl_ipums_cps(cps_months, c("EARNWT", "HOURWAGE2"))

# use dl_ipums_micro with a custom extract
extract = ipumsr::define_extract_micro(
  collection = "cps",
  description = "CPS ASEC extract",
  samples = c("cps2021_03s", "cps2022_03s", "cps2023_03s"),
  variables = c("YEAR", "OFFPOV", "ASECWT")
)
dl_ipums_micro(extract)
\dontshow{\}) # examplesIf}
}
