% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_groups.R
\name{summarize_groups}
\alias{summarize_groups}
\title{summarize_groups: summarize distinct groups}
\usage{
summarize_groups(.data, .groups, ...)
}
\arguments{
\item{...}{one or two variables, for a one- or two-way cross-tabulation}

\item{data}{a data frame}

\item{w}{weight}

\item{col}{column percentages in a two-way cross-tabulation}

\item{row}{row percentages in a two-way cross-tabulation}
}
\value{
a tibble
}
\description{
summarize_groups: summarize distinct groups
}
\examples{
groups <- c("cyl", "gear", "carb")
summarize_groups(mtcars, groups, median(mpg), mean(hp))
}
