% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{binipolate}
\alias{binipolate}
\title{Binipolate
Calculates binned interpolated percentiles
If classical = TRUE, uses \code{\link[MetricsWeighted:weighted_quantile]{MetricsWeighted::weighted_quantile()}}}
\usage{
binipolate(data, var, probs = 0.5, bin_size, .by = NULL, w = NULL)
}
\arguments{
\item{data}{a data frame}

\item{var}{unquoted variable for analysis}

\item{probs}{percentiles to calculate}

\item{bin_size}{size of bin}

\item{w}{optional weight in data}

\item{by}{tidy selection of grouping variable}
}
\value{
a tibble
}
\description{
Binipolate
Calculates binned interpolated percentiles
If classical = TRUE, uses \code{\link[MetricsWeighted:weighted_quantile]{MetricsWeighted::weighted_quantile()}}
}
\examples{
binipolate(mtcars, var = "cyl", binsize = 0.25)
binipolate(mtcars, var = "gear", w = wt, binsize = 0.25)
binipolate(mtcars, var = "mpg", group_vars = c("cyl", "gear"), w = wt, binsize = 0.25)
}
