% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{binipolate}
\alias{binipolate}
\title{Summarize a data frame as binned interpolated percentiles}
\usage{
binipolate(data, x, probs = 0.5, bin_size, .by = NULL, w = NULL)
}
\arguments{
\item{data}{data frame}

\item{x}{column to compute}

\item{probs}{numeric vector of percentiles with values \verb{[0,1]}}

\item{bin_size}{size of binning}

\item{.by}{optional, a tidy-selection of columns for single-operation grouping}

\item{w}{numeric vector of weights the same length as x giving the weights to use for elements of x}
}
\value{
a tibble or data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This function was superseded and EPI insteads uses
\code{\link[=averaged_quantile]{averaged_quantile()}} to interpolate percentiles.
}
\examples{
binipolate(mtcars, mpg, bin_size = 0.25)
binipolate(mtcars, mpg, probs = c(0.25, 0.5, 0.75), bin_size = 0.25)
binipolate(mtcars, mpg, probs = c(0.25, 0.5, 0.75), bin_size = 0.25, .by = cyl, w = wt)
}
