% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_fred}
\alias{get_fred}
\title{Retrieve data from the FRED API}
\usage{
get_fred(series, start = NULL, end = NULL, metadata = F)
}
\arguments{
\item{series}{FRED series code}

\item{start}{Start year or date (numeric year or Date object)}

\item{end}{End year or date (numeric year or Date object)}

\item{metadata}{Flag for additional metadata}
}
\value{
A tibble
}
\description{
This function is simply a wrapper around \href{https://sboysel.github.io/fredr/}{\code{fredr}} and requires you to have an \href{https://fred.stlouisfed.org/docs/api/api_key.html}{FRED API key} saved in the \code{FRED_API_KEY} environment variable.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
get_fred("UNRATE")

series = c(
  gdp = "GDP",
  urate = "UNRATE"
)

get_fred(series, start = as.Date("2024-07-01"), end = 2024)

# Tibble with list columns including additional metadata
complete_results = get_fred(series, start = 2020, end = 2025, metadata = T)
complete_results

complete_results |>
  tidyr::unnest(metadata)
\dontshow{\}) # examplesIf}
}
