% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{interpolated_median}
\alias{interpolated_median}
\title{Calculate the binned interpolated median}
\usage{
interpolated_median(x, bin_size, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector or an R object}

\item{bin_size}{size used for binning}

\item{w}{numeric vector of weights the same length as x giving the weights to use for elements of x}

\item{na.rm}{logical; if true, any NA or NaN's are removed from x before computation}
}
\value{
numeric vector
}
\description{
#' @description
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
This function is superceded by EPI's current preferred
method for interpolating medians, \code{\link[=averaged_median]{averaged_median()}}.
}
\examples{
interpolated_median(x = mtcars$mpg, bin_size = 0.50)
}
