% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{interpolated_quantile}
\alias{interpolated_quantile}
\title{interpolated_quantile: binned interpolated quantiles}
\usage{
interpolated_quantile(x, bin_size, probs = 0.5, w = NULL, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
interpolated_quantile: binned interpolated quantiles
}
