% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolated_quantile.R
\name{interpolated_quantile_p}
\alias{interpolated_quantile_p}
\title{interpolated_quantile helper}
\usage{
interpolated_quantile_p(data, p)
}
\arguments{
\item{data}{data frame, data frame extension (e.g. a tibble), or a lazy data frame (dbplyr, dtplyr) inherited from tidyverse}

\item{p}{numeric; percentile with value \verb{[0,1]}}
}
\value{
a numeric vector of length 1
}
\description{
interpolated_quantile helper
}
