% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_groups.R
\name{summarize_groups}
\alias{summarize_groups}
\title{Summarize distinct groups}
\usage{
summarize_groups(.data, .groups, ...)
}
\arguments{
\item{.data}{a data frame}

\item{.groups}{grouping variables as a tidy selection specification
of columns, as used in \code{dplyr::select()}}

\item{...}{name-value pairs passed to dplyr::summarize()}
}
\value{
a tibble
}
\description{
Summarize distinct groups
}
\examples{
summarize_groups(mtcars, cyl|gear|carb, median(mpg), mean(hp))
}
